fixed obj parser issues. indexes are 32 bit, etc.

+++ cognitics/mesh2cdb/quickobj.h
@@ -111,9 +111,9 @@ namespace cognitics {
         std::vector<QuickVert> verts;
         std::vector<QuickVert> norms;
         std::vector<QuickVert> uvs;
-        std::vector<uint16_t> vertIdxs;
-        std::vector<uint16_t> uvIdxs;
-        std::vector<uint16_t> normIdxs;
+        std::vector<uint32_t> vertIdxs;
+        std::vector<uint32_t> uvIdxs;
+        std::vector<uint32_t> normIdxs;
         std::string objFilename;
         std::string textureFilename;
         std::string textureDirectory;
